requirements.txt
openai
python-dotenv
streamlit
openai>=1.0.0

